requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
attrs==23.1.0
cfgv==3.4.0
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
drf-spectacular==0.26.5
filelock==3.12.4
filetype==1.2.0
flake8==4.0.1
flake8-docstrings==1.6.0
identify==2.5.29
inflection==0.5.1
isort==5.10.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.6.1
mysqlclient==2.2.0
nodeenv==1.8.0
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
Pillow==10.1.0
platformdirs==3.10.0
pre-commit==2.17.0
pycodestyle==2.8.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydocstyle==6.3.0
pyflakes==2.4.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.24.5
